size_t fixes to make Jardik happy + one bugfix
git-svn-id: https://skeldal.svn.sourceforge.net/svnroot/skeldal@161 27359af5-ae43-0410-a23f-8b0a9d901db5

+++ branches/next_ghost/game/dialogs.cpp
@@ -75,7 +75,7 @@ public:
 	unsigned find(unsigned num) const;
 	int findNum(unsigned pos) const;
 	MemoryReadStream *data(void) const { return _data; }
-	unsigned memsize(void) const { return _data->memsize() + sizeof(*this); }
+	size_t memsize(void) const { return _data->memsize() + sizeof(*this); }
 };
 
 #define STR_BUFF_SIZ 4096
