+++ branches/next_ghost/game/globals.h
@@ -796,7 +796,7 @@ public:
 	void addNote(int x, int y, int depth, const char *str);
 	void removeNote(unsigned idx);
 
-	unsigned memsize(void) const;
+	size_t memsize(void) const;
 };
 
 extern Map gameMap;
@@ -1688,7 +1688,7 @@ public:
 	~PalBlock(void);
 
 	const pal_t *getPal(unsigned pal) const;
-	unsigned memsize(void) const { return _size * 2 * SHADE_STEPS * PAL_SIZE * sizeof(uint8_t) + sizeof(*this); }
+	size_t memsize(void) const { return _size * 2 * SHADE_STEPS * PAL_SIZE * sizeof(uint8_t) + sizeof(*this); }
 };
 
 void add_window(int x,int y,int xs,int ys,int texture,int border,int txtx,int txty);
